+++ CHANGELOG.md
@@ -4,6 +4,12 @@ The following is a curated list of changes in the Enact limestone module, newest
 
 ## [unreleased]
 
+### Added
+
+- `limestone/Scroller` prop `stickTo` to allow the scroller to stick to a specific element
+- `limestone/VirtualGridList` prop `stickTo` to allow the scroller to stick to a specific element
+- `limestone/VirtualList` prop `stickTo` to allow the scroller to stick to a specific element
+
 ### Changed
 
 - `limestone/Button` styling to match the latest GUI